+++ src/lib/Server/Gateways/Http.ts
@@ -14,6 +14,7 @@
  * limitations under the License.
  */
 
+import * as $Net from 'net';
 import * as C from '../Common';
 import * as G from '../../Common';
 import * as E from '../Errors';
@@ -41,8 +42,8 @@ class HttpGateway implements C.IGateway {
     ) => void;
 
     public constructor(
-        private _host: string,
-        private _port: number,
+        public host: string,
+        public port: number,
         private _backlog?: number,
     ) { }
 
@@ -132,11 +133,15 @@ class HttpGateway implements C.IGateway {
 
                 return new Promise((resolve, reject) => {
 
-                    this._server.listen(this._port, this._host, this._backlog)
+                    this._server.listen(this.port, this.host, this._backlog)
                         .once('listening', () => {
 
                             this._server.removeAllListeners('error');
                             this._status = EStatus.WORKING;
+                            if (this.port === 0) {
+
+                                this.port = (this._server.address() as $Net.AddressInfo).port;
+                            }
                             resolve();
                         })
                         .once('error', (e) => {